[All] Start a rewrite of the project
This commit completely rewrites the mod loading part of PML. It has been
simplified heavily to remove the un-needed complexity and to make the code
easier to read. Kotlin is being used in the main project now as it will
bring better code readability to the project. Things like the EventBus and
Class Bindings will remain in PufferfishAPI. Mods now require a pml-mod.json
file to be recognized by the mod loader. This makes everything a lot simpler
as we don't need to iterate through classes to see if they are an instance
of the Mod class.

Things that still need to be added, like a class transformation system,
which I will add in a later commit. At the moment, the test mod loads and
is executed fine.