[Launch] Add class transformation system
This patch adds a basic ClassTransformer interface and a PMLClassLoader
which allows anyone to transform a class after it has been registered.
This should be done as soon as your mod is initialized (Mod#initialize)
to maximise the amount of classes that you can inject.
It would preferably be done in a version specific implementation, like
in 1.8.9's PMLClientMain to ensure that it functions on all versions, as
no class name is the same between mappings. Injector will be used in
PufferfishAPI to simplify this process!