[ModLoader] Remove id field from Mod interface
This patch removes the useless ID field from the mod interface. The id is
already specified in the pml-mod.json, so there is no point in having it
in the class as well! :p

This commit also changes the `newInstance()` usage to
`getDeclaredConstructor().newInstance()` as `newInstance()` is deprecated
in newer versions of Java.